requirements.txt
streamlit==1.39.0
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
sentence-transformers==3.2.1
PyPDF2==3.0.1
spacy==3.7.5
pytesseract==0.3.10
pdf2image==1.17.0
pillow==10.4.0
transformers==4.44.2
torch>=2.2.0
python-dotenv==1.0.0
groq==0.4.0
openai==1.52.2
tqdm==4.66.5
fastapi==0.115.2
uvicorn[standard]==0.31.1
python-multipart==0.0.9

